Redesign conditional mem ops for ir_intrp
The original design relied on constant propergation
and a DCE pass that I hadn't written yet.
I was conserned I'd have to write these just to get
the ir interpeter working correctly.

Then I realsed my design was overfly complex and
this approach of just putting a conditional input to
the memory operation is much smarter.