Sketch out MMIO registers
We actually have quite a bit of the NES registers defined.

But next we need to solve the "how memory is implemented" problem.
In this commit, we are upto ppuData, which needs to bridge across to
another memory bus.

I've been thinking and I suspect the solution is to repersent memery
buses as IR functions, which the interpeter can simply call to do a
memory operation, and the jit can go full-out specializing and inlining.

ppuData can be implemented as a call to the cpu_bus IR function, which
calls to the ppu_bus function.